Rename programs main app to admin
