+++ src/js/apps/admin/admin-main_app.js
@@ -1,9 +1,9 @@
 import RouterApp from 'js/base/routerapp';
 
-import ProgramsAllApp from 'js/apps/programs/list/programs-all_app';
+import ProgramsAllApp from 'js/apps/admin/list/programs-all_app';
 
 export default RouterApp.extend({
-  routerAppName: 'ProgramsApp',
+  routerAppName: 'AdminApp',
 
   childApps: {
     programsAll: ProgramsAllApp,
