+++ src/js/views/admin/list/programs-all_views.js
