+++ src/js/views/admin/list/programs-list.scss
